// angular/src/app/models/partner.model.ts
export interface Partner {
  partnerID: string;
  partnerName: string;
  location: string;
  countryID: string;
  registeredDate: string;
  spocUserID: string;
  address1: string;
  billingAddress1: string;
  partnerImage: string;
  partnerStatus: number;
  createdBy: string;
  created: string;
  lastModifiedBy: string;
  lastModified:string
}